+++ src/main/java/com/commercehub/gradle/plugin/avro/AvroExtension.java
@@ -1,5 +1,6 @@
 package com.commercehub.gradle.plugin.avro;
 
 public interface AvroExtension {
+    String getEncoding();
     String getStringType();
 }