+++ src/main/java/com/commercehub/gradle/plugin/avro/DefaultAvroExtension.java
@@ -1,7 +1,17 @@
 package com.commercehub.gradle.plugin.avro;
 
 public class DefaultAvroExtension implements AvroExtension {
+    private String encoding;
     private String stringType;
+
+    @Override
+    public String getEncoding() {
+        return encoding;
+    }
+
+    public void setEncoding(String encoding) {
+        this.encoding = encoding;
+    }
 
     @Override
     public String getStringType() {